diffmatchpatch: Fix slice memory corruption. (#1)
Before the change that introduced the use of []rune for diff, memory
operation were performed on string type and this type is immutable.
Which mean concatenation operation were actually creating a new
memory area for the result of that concatenation. But slices are
mutable and so in some situation we were corrupting the memory
area of other slices while appending some data to another one.